Refactor the Ensembl loading step to directly download data from the Ensembl public database server rather than repeatedly querying the Ensembl REST API.
scripts/download_ensembl.sh is used to perform the query (note that it needs the mysql command-line client to be installed) and create the output file. The output file is placed in mrtarget/resources.

The script should be called automatically by the Makefile when the downloaded file does not exist.

The download step takes approximately 5 seconds, depending on the speed of your Internet connection. The loading step takes approximately 30s. (The API-based approach took about 27 minutes). The new approach should also be more reliable as it removes hundreds of potentially-failing API calls.